requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
camoufox[geoip]
pytesseract
Pillow
pyngrok>=6.0.0